requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dict2xml==1.7.1
dicttoxml==1.7.4
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
idna==3.3
pytz==2022.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.9
